+++ ops/c/src/cuda/ops_cuda_decl.c
@@ -95,7 +95,7 @@ ops_dat ops_decl_dat_char(ops_block block, int size, int *dat_size, int *base,
   ops_dat dat = ops_decl_dat_temp_core(block, size, dat_size, base, d_m, d_p,
                                        stride, data, type_size, type, name);
 
-  int bytes = size * type_size;
+  size_t bytes = size * type_size;
   for (int i = 0; i < block->dims; i++)
     bytes = bytes * dat->size[i];
